Refactor Calculatot.js and Create Button.js
