+++ src/components/Calculator.js
@@ -1,10 +1,9 @@
 import React, { useState } from 'react';
 import calculate from '../logic/calculate';
+import Buttons from './Buttons';
 
 const Calculator = () => {
   const [state, setState] = useState({ total: null, next: null, operation: null });
-
-  const buttons = ['AC', '+/-', '%', '÷', '7', '8', '9', 'x', '4', '5', '6', '-', '1', '2', '3', '+', '0', '.', '='];
 
   const runCalculator = (btn) => {
     const newObj = calculate(state, btn);
@@ -29,31 +28,12 @@ const Calculator = () => {
     return 0;
   };
 
-  const generateBtnClass = (className) => {
-    if (className === '÷' || className === 'x' || className === '-'
-    || className === '+' || className === '=') {
-      return 'orange';
-    } if (className === '0') { return 'light-grey zero'; }
-    return 'light-grey';
-  };
-
   return (
-    <div className="calculator">
-      <div className="header">{displayResult() }</div>
-      {buttons.map((btn) => (
-        <button
-          type="button"
-          key={btn}
-          className={generateBtnClass(btn)}
-          onClick={() => runCalculator(btn)}
-          onKeyDown={(e) => { if (e.key === 'Enter')handleKeyDown(btn); }}
-        >
-          {' '}
-          {btn}
-          {' '}
-        </button>
-      ))}
-    </div>
+    <Buttons
+      displayResult={displayResult}
+      runCalculator={runCalculator}
+      handleKeyDown={handleKeyDown}
+    />
   );
 };
